+++ pycobertura/cli.py
@@ -4,9 +4,11 @@ from pycobertura.cobertura import Cobertura
 from pycobertura.reporters import (
     HtmlReporter,
     TextReporter,
+    MarkdownReporter,
     JsonReporter,
     HtmlReporterDelta,
     TextReporterDelta,
+    MarkdownReporterDelta,
     JsonReporterDelta,
 )
 from pycobertura.filesystem import filesystem_factory
@@ -18,6 +20,7 @@ pycobertura = click.Group()
 reporters = {
     "html": HtmlReporter,
     "text": TextReporter,
+    "markdown": MarkdownReporter,
     "json": JsonReporter,
 }
 
@@ -95,6 +98,7 @@ def show(cobertura_file, format, output, source, source_prefix):
 
 delta_reporters = {
     "text": TextReporterDelta,
+    "markdown": MarkdownReporterDelta,
     "html": HtmlReporterDelta,
     "json": JsonReporterDelta,
 }
@@ -204,7 +208,7 @@ def diff(
 
     isatty = True if output is None else output.isatty()
 
-    if format in {"text", "json"}:
+    if format in {"text", "json", "markdown"}:
         color = isatty if color is None else color is True
         reporter_kwargs["color"] = color