+++ CHANGES.md
@@ -11,8 +11,9 @@
   * do not use object in class inheritance as this is default behaviour in Python3
   * move formatting in `reporters.py` into `Reporter` and `DeltaReporter` base classes to have a uniform formatting: new `stringify` in `utils.py` method and corresponding test in `test_stringify.py`. The subclasses and only inherit from these two base classes to ensure a consistent formatting across the reporters.
   * change datastructure of coverage information from `namedtuple` (immutable) to `dictionary` (mutable) in `reporters.py`. Adjusted `html-delta.jinja2`, `html.jinja2` and `filters.py` accordingly. This change in datastructure leads to a more compact and more readable code.
-  * `pycobertura show` now includes json format
-  * `pycobertura diff` now includes json format
+  * `pycobertura show` and `pycobertura diff` now include json format
+  * `pycobertura show` and `pycobertura diff` now include markdown format
+
   Thanks @gro1m
 
 ## 2.1.0 (2021-09-23)